#135 adding circ api integration

+++ src/main/java/com/garv/satta/fantasy/external/DTO/PlayerCricDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class PlayerCricDTO {
     private String name;
-    private Long pid;
+    private Integer pid;
 }
